Implement reviewer feedback
