+++ bedrock/extract/disaggregation/export_cornerstone_waste_disagg_matrices.py
@@ -1,17 +1,14 @@
 """Export waste-disaggregated Cornerstone matrices for offline electricity disaggregation.
 
-Requires active :class:`~bedrock.utils.config.usa_config.USAConfig` to match every key
-pinned in ``2025_usa_cornerstone_full_model.yaml`` and loadable waste disaggregation
-weights. The canonical YAML itself (not a duplicated flag list) is the source of truth
-for the precondition.
+Requires ``USA_CONFIG_FILE=2025_usa_cornerstone_full_model.yaml`` and loadable waste
+disaggregation weights. Intended as a one-time offline script to generate electricity
+disaggregation inputs, not as a pipeline integration point.
 """
 
 from __future__ import annotations
 
 import os
 import pathlib
-
-import yaml
 
 from bedrock.publish.excel.writer import clear_publish_caches
 from bedrock.transform.allocation.derived import derive_E_usa
@@ -22,48 +19,21 @@ from bedrock.transform.eeio.derived_cornerstone import (
     derive_cornerstone_Ytot_full_cs_matrix,
     get_waste_disagg_weights,
 )
-from bedrock.utils.config.usa_config import CONFIG_DIR, get_usa_config
+from bedrock.utils.config.usa_config import USA_CONFIG_ENV_VAR
 
 _DISAGG_ROOT: pathlib.Path = pathlib.Path(__file__).resolve().parent
 
 _REQUIRED_CONFIG_FILE: str = "2025_usa_cornerstone_full_model.yaml"
 
 
-def _load_required_yaml_pins() -> dict[str, object]:
-    """Return the raw pinned keys from the canonical full-model YAML.
-
-    Only keys explicitly set in ``2025_usa_cornerstone_full_model.yaml`` are
-    returned; defaulted USAConfig fields are excluded so future YAML edits
-    flow through without code drift.
-    """
-    yaml_path = os.path.join(CONFIG_DIR, _REQUIRED_CONFIG_FILE)
-    with open(yaml_path) as f:
-        loaded = yaml.safe_load(f) or {}
-    if not isinstance(loaded, dict):
-        raise RuntimeError(
-            f"{_REQUIRED_CONFIG_FILE} did not parse to a mapping; got "
-            f"{type(loaded).__name__}"
-        )
-    return loaded
-
-
 def assert_cornerstone_matrix_export_preconditions() -> None:
-    """Raise ``RuntimeError`` if the active USAConfig deviates from the
-    canonical full-model YAML or if waste disaggregation weights are missing.
-    """
-    expected = _load_required_yaml_pins()
-    cfg = get_usa_config()
-    mismatches: list[str] = []
-    for key, expected_value in expected.items():
-        actual_value = getattr(cfg, key, None)
-        if actual_value != expected_value:
-            mismatches.append(f"{key}={actual_value!r} (expected {expected_value!r})")
-    if mismatches:
+    """Raise ``RuntimeError`` if the active config file or waste weights are wrong."""
+    if os.environ.get(USA_CONFIG_ENV_VAR) != _REQUIRED_CONFIG_FILE:
         raise RuntimeError(
-            "Cornerstone matrix export requires USAConfig to match "
-            f"`{_REQUIRED_CONFIG_FILE}`: "
-            + "; ".join(mismatches)
-            + f". Load that YAML via set_global_usa_config(\"{_REQUIRED_CONFIG_FILE}\") "
+            "Cornerstone matrix export requires "
+            f"{USA_CONFIG_ENV_VAR}={_REQUIRED_CONFIG_FILE!r}; got "
+            f"{os.environ.get(USA_CONFIG_ENV_VAR)!r}. "
+            f'Load that YAML via set_global_usa_config("{_REQUIRED_CONFIG_FILE}") '
             "before exporting."
         )
     if get_waste_disagg_weights() is None:
@@ -91,6 +61,8 @@ def export_cornerstone_matrices_to_csv(
     )
     out.mkdir(parents=True, exist_ok=True)
 
+    # Follows the derivation process of derived_cornerstone.py,
+    # except for derive_cornerstone_Ytot_full_cs_matrix() which recreates the full Y matrix in CS space
     V = derive_cornerstone_V()
     uset = derive_cornerstone_U_set()
     VA = derive_cornerstone_VA()
@@ -107,11 +79,7 @@ def export_cornerstone_matrices_to_csv(
 
 
 if __name__ == "__main__":
-    # Offline use: set USA_CONFIG_FILE or call set_global_usa_config first.
-    from bedrock.utils.config.usa_config import (
-        USA_CONFIG_ENV_VAR,
-        set_global_usa_config,
-    )
+    from bedrock.utils.config.usa_config import set_global_usa_config
 
     if not os.environ.get(USA_CONFIG_ENV_VAR):
         set_global_usa_config(_REQUIRED_CONFIG_FILE)
